Improve OTA Dictionary Rest controller

+++ src/main/java/cs/vsu/otamapper/service/OTADictionaryService.java
@@ -4,6 +4,10 @@ import cs.vsu.otamapper.entity.OTADictionary;
 import cs.vsu.otamapper.repository.OTADictionaryRepository;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
+import java.util.List;
+import java.util.stream.Collectors;
+
 @Service
 public class OTADictionaryService {
 
@@ -23,7 +27,22 @@ public class OTADictionaryService {
         return result;
     }
 
+    public List<OTADictionary> findAllWithoutParameters(long userId) {
+        return dictionaryRepository.getIdAndNameByUser_Id(userId).stream()
+                .map(d -> new OTADictionary((long) d[0], (String) d[1]))
+                .collect(Collectors.toList());
+    }
+
+    public OTADictionary findById(long id) {
+        return dictionaryRepository.findById(id).orElse(null);
+    }
+
+    @Transactional
     public void delete(Long id) {
-        dictionaryRepository.deleteById(id);
+        OTADictionary otaDictionary = findById(id);
+        if (otaDictionary != null) {
+            parameterService.deleteAllByDictionaryId(otaDictionary);
+            dictionaryRepository.deleteById(id);
+        }
     }
 }
